refactor(security-events): the logout profile's steps get their own folder
Five of the fifteen files in BackChannelLogout are pipeline steps, and the core already keeps its
own that way. Same shape, same reading: the folder listing now separates what the profile is made
of from what surrounds it - the intake, the notification, the sink, the validator.
